TcpInstruments.jl

Documentation for TcpInstruments.jl

TcpInstruments.jl

Index

Supported Instruments

ImpedanceAnalyzer

Oscilloscope

MultiMeter

PowerSupply

WaveformGenerator

Index

TcpInstruments.TcpInstruments

TcpInstruments.Agilent4294A

TcpInstruments.Agilent4395A

TcpInstruments.AgilentE36312A

TcpInstruments.ImpedanceAnalyzer

TcpInstruments.Instrument

TcpInstruments.Keysight33612A

TcpInstruments.KeysightDMM34465A

TcpInstruments.MultiMeter

TcpInstruments.Oscilloscope

TcpInstruments.PowerSupply

TcpInstruments.SRSPS310

TcpInstruments.VersatilePower

TcpInstruments.WaveformGenerator

TcpInstruments.XYZStage

TcpInstruments.disable_output

TcpInstruments.disable_output

TcpInstruments.disable_output

TcpInstruments.enable_output

TcpInstruments.enable_output

TcpInstruments.enable_output

TcpInstruments.f_query

TcpInstruments.get_bandwith

TcpInstruments.get_bandwith

TcpInstruments.get_burst_mode

TcpInstruments.get_burst_num_of_cycles

TcpInstruments.get_channel

TcpInstruments.get_channel

TcpInstruments.get_channel

TcpInstruments.get_channel

TcpInstruments.get_coupling

TcpInstruments.get_current

TcpInstruments.get_current_limit

TcpInstruments.get_current_limit

TcpInstruments.get_current_limit

TcpInstruments.get_impedance

TcpInstruments.get_impedance

TcpInstruments.get_lpf_state

TcpInstruments.get_mode

TcpInstruments.get_output

TcpInstruments.get_output

TcpInstruments.get_output

TcpInstruments.get_resistance

TcpInstruments.get_tc_temperature

TcpInstruments.get_temp_unit

TcpInstruments.get_volt_ac

TcpInstruments.get_voltage

TcpInstruments.get_voltage

TcpInstruments.get_voltage

TcpInstruments.get_voltage

TcpInstruments.get_voltage_limit

TcpInstruments.initialize

TcpInstruments.lpf_off

TcpInstruments.lpf_on

TcpInstruments.query

TcpInstruments.run

TcpInstruments.set_bandwith

TcpInstruments.set_bandwith

TcpInstruments.set_burst_mode_gated

TcpInstruments.set_burst_mode_trigger

TcpInstruments.set_burst_num_of_cycles

TcpInstruments.set_burst_period

TcpInstruments.set_channel

TcpInstruments.set_channel

TcpInstruments.set_channel

TcpInstruments.set_current_limit

TcpInstruments.set_current_limit

TcpInstruments.set_current_limit

TcpInstruments.set_function

TcpInstruments.set_impedance_1Mohm

TcpInstruments.set_impedance_50ohm

TcpInstruments.set_mode_burst

TcpInstruments.set_mode_cw

TcpInstruments.set_tc_type

TcpInstruments.set_temp_unit_celsius

TcpInstruments.set_temp_unit_farenheit

TcpInstruments.set_temp_unit_kelvin

TcpInstruments.set_volt_ac

TcpInstruments.set_voltage

TcpInstruments.set_voltage

TcpInstruments.set_voltage

TcpInstruments.set_voltage_limit

TcpInstruments.split_str_into_host_and_port

TcpInstruments.stop

TcpInstruments.terminate

TcpInstruments.TcpInstruments — Module

TcpInstruments is an ongoing effort to provide a simple unified interface to common lab equipment.

To connect to an instrument use:

handle = initialize({Instrument-Type}, "IP-ADDRESS:PORT")

To prevent hardcoding in scripts and/or to make it simplier for your lab or organization to keep track of all the equipment and ip addresses a configuration file can also be specified.

To use Orchard Ultrasound's latest lab config use:

create_config()

This will create a yaml file in your home directory: ~/.tcp.yml

This yaml file will be loaded everytime you use this package.

You can also create a project-specific config by creating the config in your project root directory instead of your home directory. You can do this with:

create_config(pwd())

Once you have created a config file you can change it with

edit_config()

If the instrument has its address in the config file you can now connect with:

handle = initialize({Instrument-Type})

To see the different types of devices you can interface with use help>Instrument.

TcpInstruments.Agilent4294A — Type

http://literature.cdn.keysight.com/litweb/pdf/04294-90061.pdf

Available functions

initialize()

terminate()

get_frequency_range()

`setfrequencyrange([start, stop]) # in hertz)

get_num_data_points(x)

number of points on x-axis / number of samples

set_num_data_points(x)

get_impedance() # get the data

set_volt_ac

get_volt_ac

get_volt_dc

set_volt_dc

bandwidth ({1,2,3,4,5})

1 -> lowest bandwidth, 5 -> highest bandwidth

TcpInstruments.Agilent4395A — Type

http://literature.cdn.keysight.com/litweb/pdf/04395-90031.pdf

Available functions

initialize()

terminate()

get_frequency_range()

`setfrequencyrange([start, stop]) # in hertz)

get_num_data_points(x)

number of points on x-axis / number of samples

set_num_data_points(x)

get_impedance() # get the data

set_volt_ac

get_volt_ac

get_volt_dc

set_volt_dc

bandwidth ({1,2,3,4,5})

1 -> lowest bandwidth, 35 -> highest bandwidth

TcpInstruments.AgilentE36312A — Type

Available functions

enable_output()

disable_output()

set_voltage(voltage)

get_voltage()

set_current_limit(current)

get_current_limit()

set_channel(channel)

get_channel()

TcpInstruments.ImpedanceAnalyzer — Type

Supported Instruments:

Agilent4395A

TcpInstruments.Instrument — Type

An instrument is a generic device with which you can take and read measurements

All instruments can be initialized and terminated. For more information on how to connect to a supported instrument:

help>initialize

You can use the help feature on any supported instrument group.

For more information on the actual instruments you can initialize use julia's help feature on one of the supported instrument groups.

For example lets say you want to learn about oscilloscopes.

julia>?
help>Oscilloscope

  Supported Instruments:
  ≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡

    •  AgilentDSOX4034A

Pick one of the supported instruments for more information on how to use it and for all its available functions:

help>AgilentDSOX4034A

Supported Instrument Groups

Oscilloscope

MultiMeter

PowerSupply

WaveformGenerator

ImpedanceAnalyzer

XYZStage

TcpInstruments.Keysight33612A — Type

Device has two channels: 1 & 2

Supported functions

initialize()

terminate()

enable_output()

disable_output()

get_output()

get_frequency()

set_frequency()

get_amplitude()

set_amplitude()

get_burst_num_cycles()

set_burst_num_cycles()

get_time_offset(): Not implemented

set_time_offset(): Not implemented

get_voltage_offset()

set_voltage_offset()

get_burst_period()

set_burst_period()

get_mode()

set_mode_burst(; mode=:trigger, trig_src=:timer)

set_mode_cw()

Creating a Sin Wave Example:

wave = initialize(Keysight33612A, "10.1.30.36")
set_mode_cw(wave) # Set to continuous waveform mode
set_function(wave, "SIN")
set_frequency(wave, "1e3") # Frequency val can be string or number
set_amplitude(wave, 0.1)
set_voltage_offset(wave, 0)
enable_output(wave) # Starts wave

TcpInstruments.KeysightDMM34465A — Type

Available functions

initialize

terminate

get_tc_temperature (tc = thermocouple)

get_voltage

get_current

get_resistance(;wire) # wire must be set to 2 or 4

get_channel # (some kind of input detection not selection)

TcpInstruments.MultiMeter — Type

Supported Instruments

KeysightDMM34465A

TcpInstruments.Oscilloscope — Type

Supported Instruments

AgilentDSOX4034A

TcpInstruments.PowerSupply — Type

Supported Instruments

AgilentE36312A

VersatilePower

PS310

TcpInstruments.SRSPS310 — Type

GPIB Enabled Device. Requires an attached Prologix Controller to work.

Available functions

enable_output()

disable_output()

set_voltage(volts)

get_voltage()

set_voltage_limit(volts)

get_voltage_limit()

set_current_limit(current)

get_current_limit()

set_prologix_chan(chan)

get_prologix_chan(chan)

TcpInstruments.VersatilePower — Type

This device only has one channel so channel specific functions and optional arguments are not available.

Available functions

enable_output()

disable_output()

set_voltage(volts)

get_voltage()

setcurrentlimit(current)

getcurrentlimit()

Helpers

remote_mode(): sets the device to remote mode. Automatically called on initialize

local_mode(): sets the device to local mode. Automatically called on terminate

This instrument has a remote and local mode. Some commands do not work while the device is in local mode, thus when initializing this device remote_mode is called automatically and the device is always in remote mode by default.

These functions should not be directly needed but if for some reason you need to switch modes while using the device you can use local_mode to turn the device back to local mode.

TcpInstruments.WaveformGenerator — Type

Supported Instruments:

Keysight33612A: Default ip ~ "10.1.30.36"

TcpInstruments.XYZStage — Type

Supported Instruments:

ThorlabsLTS150

TcpInstruments.disable_output — Method

This will disable an output on a device.

If the device has multiple channels it will disable the device for the currently selected channel. To see the channel that will effected use the get_channel function.

If you want to disable a different channel, first use set_channel to choose the channel. Running this function subsequently will disable that channel

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.disable_output — Method

This will disable an output on a device.

Arguments:
  - obj
    - must be a Power Supply Instrument
Supported Instruments:
   - Power supply

TcpInstruments.disable_output — Method

This will disable an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.enable_output — Method

This will enable an output on a device.

If the device has multiple channels it will enable the device for the currently selected channel. To see the channel that will effected use the get_channel function.

If you want to enable a different channel, first use set_channel to choose the channel. Running this function subsequently will enable that channel

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.enable_output — Method

This will enable an output on a device.

Arguments:
  - obj
    - must be a Power Supply Instrument
Supported Instruments:
   - Power supply

TcpInstruments.enable_output — Method

This will enable an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.f_query — Method

Writes a message to a device then listens for and returns any output from the device.

This is a blocking procedure and will block until a response is received from the device or till it has been blocking for longer than the designated timeout time after which an error will be thrown.

Arguments

instr::Instrument: Any instrument that supports being written to and read from

message::AbstractString: The message to be sent to the device before listening for a response

timeout: Optional flag ~ How long to try and listen for a response before giving up and throwing an error. The default time is half a second. Note: if timeout is set to 0 then this will turn off the timeout functionality and query may listen/block indefinitely for a response

Differs from query in that it will return a Float64 and not a String

TcpInstruments.get_bandwith — Method

Returns bandwidth level (1-5)

TcpInstruments.get_bandwith — Method

Returns device bandwidth

TcpInstruments.get_burst_mode — Method

Returns the burst mode of a device:
    "TRIG" ~ If the device is in Triggered Mode
    "GAT" ~ If the device is in Gated Mode

TcpInstruments.get_burst_num_of_cycles — Method

Returns number of cycles burst mode is set to

TcpInstruments.get_channel — Method

This will return the global or default channel of a device.

Allows you to see what the global channel is set to at the
moment

Parameters:
  - obj
    - must be a Power Supply Instrument
  - chan
    - This can be a string or int: 1, 2, 3 .. to n
    where n is the total number of channels
  - v
    - optional boolean flag argument that is set to false
    - setting to `true` will print the verbose channel name

Supported Instruments:
   - Power supply

Returns:
  String
    - {"1"|"2"|...}
    - when v == true: "P6V", .. etc

TcpInstruments.get_channel — Method

Returns 1 or 2 depending on current channel

TcpInstruments.get_channel — Method

This will return the global or default channel of a device.

Allows you to see what the global channel is set to at the moment

Parameters:

obj

must be a Power Supply Instrument

chan

This can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

v

optional boolean flag argument that is set to false

setting to true will print the verbose channel name

Supported Instruments:

Power supply

Returns: String - {"1"|"2"|...} - when v == true: "P6V", .. etc

TcpInstruments.get_channel — Method

Indicates which input terminals are selected on the front panel Front/Rear switch. This switch is not programmable; this query reports the position of the switch, but cannot change it.

Do not toggle the Front/Rear switch with active signals on the terminals. This switch is not intended to be used in this way and may be damaged by high voltages or currents, possibly compromising the instrument's safety features.

Returns

"FRON" or "REAR"

TcpInstruments.get_coupling — Method

get_coupling(scope, chan=1)

returns "AC" or "DC"

TcpInstruments.get_current — Method

Returns current

Arguments

type: "DC" | "AC"

TcpInstruments.get_current_limit — Method

This will return the current limit of a device.

Parameters:

obj

must be a Power Supply Instrument

chan

This is an optional parameter

If not provided it will use the default channel (see set_channel)

Otherwise this can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Current Limit

TcpInstruments.get_current_limit — Method

This will return the current limit of a device.

Supported Instruments:
   - Power supply

Returns:
  Current Limit

TcpInstruments.get_current_limit — Method

This will return the current limit of a device.

Supported Instruments:

Power supply

Returns: Current Limit

TcpInstruments.get_impedance — Method

P. 188 Must be in ZA mode IMPH COMP

TcpInstruments.get_impedance — Method

P. 3-10

Error corrected data The results of error correction are stored in the data arrays as complex number pairs. Formatted data

This is the array of data being displayed. It reflects all post-processing functions such as electrical delay, and the units of the array read out depends on the current display format.

TcpInstruments.get_lpf_state — Method

get_lpf_state(scope, chan=1)

See state the internal low-pass filter:

returns "0" or "1"

TcpInstruments.get_mode — Method

Returns: "CW" ~ if device is in continous wavefrom mode "BURST" ~ if device is in BURST mode

TcpInstruments.get_output — Method

This will return the state of an output on a device.

If the device has multiple channels is will display the state of the currently selected channel. To see the channel that will read use the get_channel function.

If you want to see the state of a different channel, first use set_channel to choose the channel. Running this function subsequently will disable that channel

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

Returns: String: {"0"|"1"}

TcpInstruments.get_output — Method

This will disable an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

Returns

true if High Voltage Output is Off

false if High Voltage Output is On

TcpInstruments.get_output — Method

This will return the state of an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

Returns: String: {"OFF"|"ON"}

TcpInstruments.get_resistance — Method

Returns resistance

Arguments

type: "DC" | "AC"

TcpInstruments.get_tc_temperature — Method

get_tc_temperature(multimeter)

Page. 288

Perform take a measurement with the probe mode set to thermocouple

TcpInstruments.get_temp_unit — Method

get_temp_unit(multimeter)

Returns C, F or K depending on the set temperature unit

TcpInstruments.get_volt_ac — Method

Returns oscillator (ac) voltage

TcpInstruments.get_voltage — Method

This will return the voltage of a device's channel.

Parameters:

obj

must be a Power Supply Instrument

chan

This is an optional parameter

If not provided it will use the default channel (see set_channel)

Otherwise this can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Voltage

TcpInstruments.get_voltage — Method

Returns voltage

Arguments

type: "DC" | "AC"

TcpInstruments.get_voltage — Method

This will return the voltage of a device

Voltage Limit: 1250V

Supported Instruments:
   - Power supply

Returns:
  Voltage

TcpInstruments.get_voltage — Method

This will return the voltage of a device

Supported Instruments:

Power supply

Returns: Voltage

TcpInstruments.get_voltage_limit — Method

This will return the voltage limit of a device

Voltage Limit: 1250V

Supported Instruments:
   - Power supply

Returns:
  Voltage

TcpInstruments.initialize — Method

initialize(model, address)

Initializes a connection to the instrument at the given (input) IP address.

Arguments

model: They device type you are connecting to. Use help>Instrument to see available options.

address::String: The ip address of the device. Ex. "10.3.30.23"

TcpInstruments.lpf_off — Function

lpf_off(scope, chan=1)

Turn off an internal low-pass filter.

TcpInstruments.lpf_on — Function

lpf_on(scope, chan=1)

Turn on an internal low-pass filter. When the filter is on, the bandwidth of the specified channel is limited to approximately 25 MHz.

TcpInstruments.query — Method

Writes a message to a device then listens for and returns any output from the device.

This is a blocking procedure and will block until a response is received from the device or till it has been blocking for longer than the designated timeout time after which an error will be thrown.

Arguments

instr::Instrument: Any instrument that supports being written to and read from

message::AbstractString: The message to be sent to the device before listening for a response

timeout: Optional flag ~ How long to try and listen for a response before giving up and throwing an error. The default time is half a second. Note: if timeout is set to 0 then this will turn off the timeout functionality and query may listen/block indefinitely for a response

TcpInstruments.run — Method

run(scope)

Run Oscilloscope

TcpInstruments.set_bandwith — Method

Pg. 274

(Initial value) Specifies bandwidth 1 (shortest measurement time)

Specifies bandwidth 2

Specifies bandwidth 3

Specifies bandwidth 4

Specifies bandwidth 5 (longest measurement time, accurate

measurement).

TcpInstruments.set_bandwith — Method

Pg. B-3

2, 10, 30, 100, 300, 1000 (=1k), 3000 (=3k), 10000 (=10k), 30000 (=30k) (Network and impedance analyzers)

TcpInstruments.set_burst_mode_gated — Method

Sets the burst mode of a device to Gated Mode

TcpInstruments.set_burst_mode_trigger — Method

Sets the burst mode of a device to Triggered Mode

TcpInstruments.set_burst_num_of_cycles — Method

Sets the number of cycles for burst mode

TcpInstruments.set_burst_period — Method

chan=<1|2>

num = {<seconds>|MINimum|MAXimum}

TcpInstruments.set_channel — Method

This will set the global channel on a device.

Any commands like set_voltage that affect the
device with respect to a specific channel will be impacted
by this command.

By setting the channel on a device, all subsequent commands
will operate on that channel unless they allow for an optional channel argument.

Parameters:
  - obj
    - must be a Power Supply Instrument
  - chan
    - This can be a string or int: 1, 2, 3 .. to n
    where n is the total number of channels

Supported Instruments:
   - Power supply

Returns:
  Nothing

TcpInstruments.set_channel — Method

set_channel(i, channel_number)

Uses.

set_channel(i, 1)

set_channel(i, 2)

TcpInstruments.set_channel — Method

This will set the global channel on a device.

Any commands like set_voltage that affect the device with respect to a specific channel will be impacted by this command.

By setting the channel on a device, all subsequent commands will operate on that channel unless they allow for an optional channel argument.

Parameters:

obj

must be a Power Supply Instrument

chan

This can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_current_limit — Method

This will change the current limit of a device on a given channel

Parameters:

obj

must be a Power Supply Instrument

num

integer or decimal of the desired current limit

chan

This is an optional parameter

If not provided it will use the default channel (see set_channel)

Otherwise this can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_current_limit — Method

This will change the current limit of a device 

MIN Value: 0
Max Value: { 2.1e-3 | 0.021 } (21mA)

Supported Instruments:
   - Power supply

Returns:
  Nothing

TcpInstruments.set_current_limit — Method

This will change the current limit of a device

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_function — Method

Acceptable inputs {SINusoid|SQUare|TRIangle|RAMP|PULSe|PRBS|NOISe|ARB|DC}

TcpInstruments.set_impedance_1Mohm — Method

set_impedance_1Mohm(scope, chan=1)

Set impedance to 1MΩ

TcpInstruments.set_impedance_50ohm — Method

set_impedance_50ohm(scope, chan=1)

Set impedance to 50Ω

TcpInstruments.set_mode_burst — Method

    set_mode_burst(instr)

Changes Waveform Generator mode from continous waveform to burst

By default it sets the type of burst mode to triggered mode. Gated mode can also be set by using the optional flag:

    set_mode_burst(instr; mode=:gated)

The optional trig_src flag sets the trigger source for burst triggered mode. Right now the default is Timer. To implement more trigger sources see page 130 of the manual for 33612A

TcpInstruments.set_mode_cw — Method

Puts the device in continuous waveform/turns off burst mode

TcpInstruments.set_tc_type — Method

set_tc_type(multimeter; type="K")

Arguments

type: Can be E, J, K, N, R, T (Defaults to K)

TcpInstruments.set_temp_unit_celsius — Method

set_temp_unit_celsius(multimeter)

Sets the temperature unit to celcius

TcpInstruments.set_temp_unit_farenheit — Method

set_temp_unit_farenheit(multimeter)

Sets the temperature unit to farenheit

TcpInstruments.set_temp_unit_kelvin — Method

set_temp_unit_kelvin(multimeter)

Sets the temperature unit to kelvin

TcpInstruments.set_volt_ac — Method

Range For voltage setting: 5E-3 to 1

TcpInstruments.set_voltage — Method

This will change the voltage output voltage of a device.

Parameters:

obj

must be a Power Supply Instrument

num

integer or decimal of the desired voltage

chan

This is an optional parameter

If not provided it will use the default channel (see set_channel)

Otherwise this can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_voltage — Method

This will change the voltage output of a device.

Voltage Limit: The value of getvoltagelimit()

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_voltage — Method

This will change the voltage output of a device.

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_voltage_limit — Method

set_voltage_limit(::SRSPS310, voltage_limit)

This will change the voltage limit of a device.

Max Voltage Limit: 1250V

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.split_str_into_host_and_port — Method

split_str_into_host_and_port(str)

Splits a string like "192.168.1.1:5056" into ("192.168.1.1", 5056)

TcpInstruments.stop — Method

stop(scope)

Stop Oscilloscope

TcpInstruments.terminate — Method

terminate(instr)

Closes the TCP connection.