General Usage

To connect to an instrument you must know its model and IP address:

using TcpInstruments
instrument_handle = initialize(AgilentDSOX4034A, "10.1.30.32")
info(instrument_handle)
data = get_data(instrument_handle)
save(data)

When you are done you can close your connection:

terminate(instrument_handle)

If you do not know the ip address of any devices on your network:

scan_network()

By default scan_network will check the addresses between 10.1.30.1 to 10.1.30.255 but you may need to scan a different range, say: 10.1.150.1-255

scan_network(; ip_network="10.1.150.")

General functions

TcpInstruments.initialize — Function

initialize(model, address)

Initializes a connection to the instrument at the given (input) IP address.

Arguments

model: They device type you are connecting to. Use help> -->Instrument to see available options.

address::String: The ip address of the device. Ex. "10.3.30.23"

Keywords

GPIB_ID::Int: The GPIB interface ID of your device. This is optional and doesn't need to be set unless you are using a prologix controller to control it remotely.

TcpInstruments.terminate — Function

terminate(instr)

Closes the TCP connection.

TcpInstruments.info — Function

info(instr_h)

Asks an instrument to print model number and other device info.

TcpInstruments.scan_network — Function

scan_network(; ip_network="10.1.30.", ip_range=1:255, v=false)

Will scan your network and report all found devices.

By default it only searches for devices connected on port: 5025

If you would like to search for devices on a different port set the v flag to true.

TcpInstruments.save — Function

save(data)
save(data; format=:matlab)
save(data; filename="custom_file_name.ext")

Save data to a file

By default saves to julia format (.jld2) but can also export data to matlab by using the format=:matlab keyword argument

TcpInstruments.load — Function

data = load("file.jld2")

Loads saved data from a file

Power Supply

AgilentE36312A

TcpInstruments.disable_output — Method

This will disable an output on a device.

If the device has multiple channels it will disable the device for the currently selected channel. To see the channel that will effected use the get_channel function.

If you want to disable a different channel, first use set_channel to choose the channel. Running this function subsequently will disable that channel

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.enable_output — Method

This will enable an output on a device.

If the device has multiple channels it will enable the device for the currently selected channel. To see the channel that will effected use the get_channel function.

If you want to enable a different channel, first use set_channel to choose the channel. Running this function subsequently will enable that channel

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.get_channel — Method

This will return the global or default channel of a device.

Allows you to see what the global channel is set to at the moment

Parameters:

obj

must be a Power Supply Instrument

chan

This can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

v

optional boolean flag argument that is set to false

setting to true will print the verbose channel name

Supported Instruments:

Power supply

Returns: String - {"1"|"2"|...} - when v == true: "P6V", .. etc

TcpInstruments.get_current_limit — Method

This will return the current limit of a device.

Parameters:

obj

must be a Power Supply Instrument

chan

This is an optional parameter

If not provided it will use the default channel (see set_channel)

Otherwise this can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Current Limit

TcpInstruments.get_output — Method

This will return the state of an output on a device.

If the device has multiple channels is will display the state of the currently selected channel. To see the channel that will read use the get_channel function.

If you want to see the state of a different channel, first use set_channel to choose the channel. Running this function subsequently will disable that channel

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

Returns: String: {"0"|"1"}

TcpInstruments.get_voltage — Method

This will return the voltage of a device's channel.

Parameters:

obj

must be a Power Supply Instrument

chan

This is an optional parameter

If not provided it will use the default channel (see set_channel)

Otherwise this can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Voltage

TcpInstruments.set_channel — Method

This will set the global channel on a device.

Any commands like set_voltage that affect the device with respect to a specific channel will be impacted by this command.

By setting the channel on a device, all subsequent commands will operate on that channel unless they allow for an optional channel argument.

Parameters:

obj

must be a Power Supply Instrument

chan

This can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_current_limit — Method

This will change the current limit of a device on a given channel

Parameters:

obj

must be a Power Supply Instrument

num

integer or decimal of the desired current limit

chan

This is an optional parameter

If not provided it will use the default channel (see set_channel)

Otherwise this can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_voltage — Method

This will change the voltage output voltage of a device.

Parameters:

obj

must be a Power Supply Instrument

num

integer or decimal of the desired voltage

chan

This is an optional parameter

If not provided it will use the default channel (see set_channel)

Otherwise this can be a string or int: 1, 2, 3 .. to n

where n is the total number of channels

Supported Instruments:

Power supply

Returns: Nothing

SRSPS310

This device needs a gpib adapter

As of right now the prologix adapter interface is the only supported adapter.

Connect your power supply to a prologix adapter then initialize using the GPIB_ID keyword argument.

If you do not know the GPIB Channel ID you can initialize your device without that flag. Then run scan_prologix on your device. This will tell you what channel is connected then manually you can use the set_prologix function to set the channel.

TcpInstruments.disable_output — Method

This will disable an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.enable_output — Method

This will enable an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.get_current_limit — Method

This will return the current limit of a device.

Supported Instruments:

Power supply

Returns: Current Limit

TcpInstruments.get_output — Method

This will disable an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

Returns

true if High Voltage Output is Off

false if High Voltage Output is On

TcpInstruments.get_voltage — Method

This will return the voltage of a device

Voltage Limit: 1250V

Supported Instruments:

Power supply

Returns: Voltage

TcpInstruments.get_voltage_limit — Method

This will return the voltage limit of a device

Voltage Limit: 1250V

Supported Instruments:

Power supply

Returns: Voltage

TcpInstruments.set_current_limit — Method

This will change the current limit of a device

MIN Value: 0 Max Value: { 2.1e-3 | 0.021 } (21mA)

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_voltage — Method

This will change the voltage output of a device.

Voltage Limit: The value of getvoltagelimit()

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_voltage_limit — Method

set_voltage_limit(::SRSPS310, voltage_limit)

This will change the voltage limit of a device.

Max Voltage Limit: 1250V

Supported Instruments:

Power supply

Returns: Nothing

VersatilePower

TcpInstruments.disable_output — Method

This will disable an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.enable_output — Method

This will enable an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

TcpInstruments.get_current_limit — Method

This will return the current limit of a device.

Supported Instruments:

Power supply

Returns: Current Limit

TcpInstruments.get_output — Method

This will return the state of an output on a device.

Arguments:

obj

must be a Power Supply Instrument

Supported Instruments:

Power supply

Returns: String: {"OFF"|"ON"}

TcpInstruments.get_voltage — Method

This will return the voltage of a device

Supported Instruments:

Power supply

Returns: Voltage

TcpInstruments.set_current_limit — Method

This will change the current limit of a device

Supported Instruments:

Power supply

Returns: Nothing

TcpInstruments.set_voltage — Method

This will change the voltage output of a device.

Supported Instruments:

Power supply

Returns: Nothing

Signal Generator

Keysight33612A

TcpInstruments.get_burst_mode — Method

get_burst_mode(instr)
get_burst_mode(instr; chan=1)

Returns the burst mode of a device: "TRIG" ~ If the device is in Triggered Mode "GAT" ~ If the device is in Gated Mode

TcpInstruments.get_burst_num_cycles — Method

get_burst_num_of_cycles(instr)
get_burst_num_of_cycles(instr; chan=2)

Keywords

chan: Specify channel: Default is 1

Returns

Float64: number of cycles burst mode is set to

TcpInstruments.get_burst_period — Method

get_burst_period(instr)
get_burst_period(instr; chan=2)

Keywords

chan: Specify channel: Default is 1

Returns

Float64: number of cycles burst mode is set to

TcpInstruments.get_function — Method

get_function(instr)
get_function(instr; chan=2)

Keywords

chan: Specify channel: Default is 1

Returns

String: Will return one of these shortened forms:

{SINusoid|SQUare|TRIangle|RAMP|PULSe|PRBS|NOISe|ARB|DC}

TcpInstruments.get_mode — Method

get_mode(instr)
get_mode(instr; chan=2)

Keywords

chan: Specify channel: Default is 1

Returns: "CW" ~ if device is in continous wavefrom mode "BURST" ~ if device is in BURST mode

TcpInstruments.set_burst_mode_gated — Method

Sets the burst mode of a device to Gated Mode

TcpInstruments.set_burst_mode_trigger — Method

Sets the burst mode of a device to Triggered Mode

TcpInstruments.set_burst_num_cycles — Method

set_burst_num_of_cycles(instr, cycles)
set_burst_num_of_cycles(instr, cycles; chan=2)

Sets the number of cycles for burst mode

Arguments

cycles

Keywords

chan: Specify channel: Default is 1

TcpInstruments.set_burst_period — Method

set_burst_period(obj, num; chan=1)

Arguments

num: The number of seconds (This value can also be "MIN" or "MAX")

Keywords

chan: Specify channel: Default is 1

TcpInstruments.set_function — Method

set_function(instr, func; chan=1)

Arguments

func::String: Acceptable inputs:

{SINusoid|SQUare|TRIangle|RAMP|PULSe|PRBS|NOISe|ARB|DC}

Keywords

chan: Specify channel: Default is 1

TcpInstruments.set_mode_burst — Method

    set_mode_burst(instr)

Changes Waveform Generator mode from continous waveform to burst

By default it sets the type of burst mode to triggered mode. Gated mode can also be set by using the optional flag:

    set_mode_burst(instr; mode=:gated)

The optional trig_src flag sets the trigger source for burst triggered mode. Right now the default is Timer. To implement more trigger sources see page 130 of the manual for 33612A

TcpInstruments.set_mode_cw — Method

setmodecw(instr) setmodecw(instr; chan=2)

Puts the device in continuous waveform/turns off burst mode

Keywords

chan: Specify channel: Default is 1

Impedance Analyzer

Agilent4294A

TcpInstruments.get_bandwidth — Method

get_bandwidth(instr)

Returns bandwidth level (1-5)

Returns

Int: Between 1 and 5

(Initial value) Specifies bandwidth 1 (shortest measurement time)

Specifies bandwidth 2

Specifies bandwidth 3

Specifies bandwidth 4

Specifies bandwidth 5 (longest measurement time, accurate

measurement).

TcpInstruments.get_volt_ac — Method

get_volt_ac(instr)

Returns oscillator (ac) voltage

TcpInstruments.set_bandwidth — Method

set_bandwith(instr, n)

Sets bandwidth level (1-5)

Arguments

n::Int: Desired bandwidth level (between 1 and 5)

TcpInstruments.set_volt_ac — Method

set_volt_ac(instr, voltage)

Arguments

voltage: Desired voltage, range for voltage setting: 5E-3 to 1

Agilent4395A

WIP

TcpInstruments.get_bandwith — Method

Returns device bandwidth

TcpInstruments.get_channel — Method

Returns 1 or 2 depending on current channel

TcpInstruments.get_impedance — Method

P. 3-10

Error corrected data The results of error correction are stored in the data arrays as complex number pairs. Formatted data

This is the array of data being displayed. It reflects all post-processing functions such as electrical delay, and the units of the array read out depends on the current display format.

TcpInstruments.set_bandwith — Method

Pg. B-3

2, 10, 30, 100, 300, 1000 (=1k), 3000 (=3k), 10000 (=10k), 30000 (=30k) (Network and impedance analyzers)

TcpInstruments.set_channel — Method

set_channel(i, channel_number)

Uses.

set_channel(i, 1)

set_channel(i, 2)

Oscilloscope

AgilentDSOX4024A

AgilentDSOX4034A

TcpInstruments.get_coupling — Method

get_coupling(scope, chan=1)

returns "AC" or "DC"

TcpInstruments.get_impedance — Method

get_impedance(scope)
set_impedance(scope, chan=2)

Keywords

chan: Specify channel: Default is 1

Returns

"FIFT": 50Ω

"ONEM": 1MΩ

TcpInstruments.get_lpf_state — Method

get_lpf_state(scope, chan=1)

See state the internal low-pass filter:

returns "0" or "1"

TcpInstruments.lpf_off — Function

lpf_off(scope, chan=1)

Turn off an internal low-pass filter.

TcpInstruments.lpf_on — Function

lpf_on(scope, chan=1)

Turn on an internal low-pass filter. When the filter is on, the bandwidth of the specified channel is limited to approximately 25 MHz.

TcpInstruments.run — Method

run(scope)

Run Oscilloscope

TcpInstruments.set_impedance_1Mohm — Method

set_impedance_1Mohm(scope, chan=1)

Set impedance to 1MΩ

TcpInstruments.set_impedance_50ohm — Method

set_impedance_50ohm(scope)
set_impedance_50ohm(scope, chan=2)

Keywords

chan: Specify channel: Default is 1

Set impedance to 50Ω

TcpInstruments.stop — Method

stop(scope)

Stop Oscilloscope

Multimeter

KeysightDMM34465A

TcpInstruments.get_channel — Method

Indicates which input terminals are selected on the front panel Front/Rear switch. This switch is not programmable; this query reports the position of the switch, but cannot change it.

Do not toggle the Front/Rear switch with active signals on the terminals. This switch is not intended to be used in this way and may be damaged by high voltages or currents, possibly compromising the instrument's safety features.

Returns

"FRON" or "REAR"

TcpInstruments.get_current — Method

Returns current

Keywords

type: "DC" | "AC" (Default DC)

TcpInstruments.get_resistance — Method

get_resistance(multimeter; wire=2)
get_resistance(multimeter; wire=4)

Returns resistance

Keywords

wire: 2 | 4 (Required)

TcpInstruments.get_tc_temperature — Method

get_tc_temperature(multimeter)

Perform take a measurement with the probe mode set to thermocouple

TcpInstruments.get_temp_unit — Method

get_temp_unit(multimeter)

Returns C, F or K depending on the set temperature unit

TcpInstruments.get_voltage — Method

Returns voltage

Keywords

type: "DC" | "AC" (Default DC)

TcpInstruments.set_tc_type — Method

set_tc_type(multimeter; type="K")

Keywords

type: Can be E, J, K, N, R, T (Defaults to K)

TcpInstruments.set_temp_unit_celsius — Method

set_temp_unit_celsius(multimeter)

Sets the temperature unit to celcius

TcpInstruments.set_temp_unit_farenheit — Method

set_temp_unit_farenheit(multimeter)

Sets the temperature unit to farenheit

TcpInstruments.set_temp_unit_kelvin — Method

set_temp_unit_kelvin(multimeter)

Sets the temperature unit to kelvin