General functions

initialize

terminate

info

scan_network

save

load

InstrumentConfig.initialize — Function

initialize(model::Type{Instrument})
initialize(model::Type{Instrument}, address::String; GPIB_ID::Int=-1)

Initializes a connection to the instrument at the given (input) IP address.

Arguments

model: The device type you are connecting to. Use help> Instrument to see available options

address (optional): The ip address of the device. Ex. "10.3.30.23". If not provided, TcpInstruments will look for the address in the config file

Keywords

GPIB_ID: The GPIB interface ID of your device. This is optional and doesn't need to be set unless you are using a prologix controller to control it remotely

InstrumentConfig.terminate — Function

terminate(instr::Instrument)

Closes the TCP connection.

TcpInstruments.info — Function

info(instr::Instrument)

Asks an instrument to print model number and other device info.

TcpInstruments.scan_network — Function

scan_network(; network_id="10.1.30.0", host_range=1:255)

By default, report all found devices between addresses 10.1.30.1 to 10.1.30.255.

Searches for devices connected on port:

5025 (scpi)

1234 (prologix)

TcpInstruments.save — Function

save(data)
save(data; format=:matlab)
save(data; filename="custom_file_name.ext")

Save data to a file

By default saves to julia format (.jld2) but can also export data to matlab by using the format=:matlab keyword argument

TcpInstruments.load — Function

data = load("file.jld2")

Loads saved data from a file